Migrate from Remix to Vite (#9)
Tried working with Remix framework since other projects used it and where quite satisfied so we got some knowledge in our company which would come in handy for code reviews and further development. Since this is kind of a prototype or at least an alpha version of the product it might be somewhen, I do not want to spent time caring about the operations but just want to throw it out there. That's why I decided to host it with GitHub Pages. Turns out Remix is a full-stack framework which distributes a production ready web-server instead of a static page which does not fit to what GitHub Pages needs.

Replace Remix with Vite + React instead. Vite is able to generate static sites and using React could lead to possible reuse of components and knowledge since Remix is based on React as well.