Move images to resources directory
To fix the warning:

Assets in public directory cannot be imported from JavaScript.
If you intend to import that asset, put the file in the src directory, and use /src/img/bund-logo.png instead of /public/img/bund-logo.png.
If you intend to use the URL of that asset, use /img/bund-logo.png?url.